+++ app/src/main/java/woowacourse/movie/ui/seat/presenter/BookingSeatPresenter.kt
@@ -1,9 +1,10 @@
-package woowacourse.movie.ui.seat
+package woowacourse.movie.ui.seat.presenter
 
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.domain.model.Seat
 import woowacourse.movie.domain.model.Seats
 import woowacourse.movie.domain.model.Theater
+import woowacourse.movie.ui.seat.contract.BookingSeatContract
 
 class BookingSeatPresenter(
     private val bookingSeatView: BookingSeatContract.View,
