+++ app/src/main/res/layout/activity_booking_complete.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.complete.BookingCompleteActivity">
+    tools:context=".ui.complete.view.BookingCompleteActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
