+++ app/src/test/java/woowacourse/movie/presenter/BookingSeatPresenterTest.kt
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.domain.model.Seat
-import woowacourse.movie.ui.seat.BookingSeatContract
-import woowacourse.movie.ui.seat.BookingSeatPresenter
+import woowacourse.movie.ui.seat.contract.BookingSeatContract
+import woowacourse.movie.ui.seat.presenter.BookingSeatPresenter
 
 class BookingSeatPresenterTest {
     private lateinit var view: BookingSeatContract.View
